Refactor purchasing flow and handle exceptions in REST API
The commit involves refactoring some parts of the code in the Cinema Room REST Service. Instead of returning an error message in the form of a ResponseEntity, the code now throws a SeatException when a user attempts to purchase an invalid seat or a seat that has already been purchased. A new class SeatException was created to handle this type of exceptions.

Moreover, a new GlobalExceptionHandler class is added to handle these exceptions and return a meaningful error to the API user. This provides a more maintainable way of handling and returning HTTP status and error message.

The tests were also simplified without changing the test scenarios to correspond to these changes.